+++ ai-saas/src/modules/meetings/ui/components/meetings-search-bar.tsx
